requirements.txt
asgiref==3.5.0
Django==3.2.7
djangorestframework==3.12.4
pytz==2022.1
sqlparse==0.4.2
